// back-end/src/member/dtos/create-member.dto.ts
export class CreateMemberDTO {
    name: string;

    sReward: number

    sCommission: number

    a1Reward: number

    a1Commission: number;

    b1Reward: number

    b1Commission: number;

    c1Reward: number

    c1Commission: number;

    aReward: number

    aCommission: number;

    bReward: number

    bCommission: number;

    cReward: number

    cCommission: number;

    bigReward: number

    bigCommission: number;

    nReward: number

    nCommission: number;
}